Refactor parsing logic
• Extract parser state into its own instance
• Every type should be able to parse its own structure
