+++ lib/Sabberworm/CSS/Value/URL.php
@@ -2,6 +2,7 @@
 
 namespace Sabberworm\CSS\Value;
 
+use Sabberworm\CSS\Parsing\ParserState;
 
 class URL extends PrimitiveValue {
 
@@ -11,6 +12,23 @@ class URL extends PrimitiveValue {
 		parent::__construct($iLineNo);
 		$this->oURL = $oURL;
 	}
+
+	public static function parse(ParserState $oParserState) {
+		$bUseUrl = $oParserState->comes('url', true);
+		if ($bUseUrl) {
+			$oParserState->consume('url');
+			$oParserState->consumeWhiteSpace();
+			$oParserState->consume('(');
+		}
+		$oParserState->consumeWhiteSpace();
+		$oResult = new URL(CSSString::parse($oParserState), $oParserState->currentLine());
+		if ($bUseUrl) {
+			$oParserState->consumeWhiteSpace();
+			$oParserState->consume(')');
+		}
+		return $oResult;
+	}
+
 
 	public function setURL(CSSString $oURL) {
 		$this->oURL = $oURL;
